test(provisionPool): extract makeFakeBankKit
